requirements.txt
aeidon==1.10.1
alabaster==0.7.16
apt-clone==0.2.1
apt-xapian-index==0.49
apturl==0.5.2
asarPy==1.0.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
Babel==2.15.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
cclib==1.6.2
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
colour==0.1.5
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cssbeautifier==1.15.1
cupshelpers==1.0
daemonize==2.5.0
dbus-python==1.2.18
defer==1.0.6
deprecation==2.1.0
diff-match-patch==20200713
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.6
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-docs==0.3.3
django-email-verification==0.3.3
django-guest-user==0.5.5
django-import-export==4.1.1
django-redis==5.4.0
django-rq==2.10.2
Django-Verify-Email==2.0.3
djangorestframework==3.15.2
djlint==1.34.1
docutils==0.20.1
EditorConfig==0.12.4
eyeD3==0.8.10
ffmpeg==1.4
filelock==3.6.0
fuse-python==1.0.2
greenlet==1.1.2
html-tag-names==0.1.2
html-void-elements==0.1.0
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.7
ifaddr==0.1.7
imagesize==1.4.1
IMDbPY==2021.4.18
importlib-metadata==4.6.4
incremental==21.3.0
iniparse==0.4
jeepney==0.7.1
Jinja2==3.1.4
jsbeautifier==1.15.1
json5==0.9.25
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
more-itertools==8.10.0
msgpack==1.0.3
mutagen==1.45.1
nemo-emblems==6.0.1
netaddr==0.8.0
netifaces==0.11.0
notify2==0.3
numpy==1.21.5
oauthlib==3.2.0
onboard==1.4.1
openrazer==3.8.0
openrazer-daemon==3.8.0
packaging==24.1
PAM==0.4.2
pathspec==0.12.1
pexpect==4.8.0
phply==1.2.5
Pillow==9.0.1
ply==3.11
protobuf==3.12.4
psutil==5.9.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycountry==20.7.3
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
pyenchant==3.2.0
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pylibacl==0.6.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyparted==3.11.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-Levenshtein==0.12.2
python-magic==0.4.24
python-xlib==0.29
pytz==2022.1
pyudev==0.22.0
pyxattr==0.7.2
pyxdg==0.27
PyYAML==6.0.1
qrcode==7.3.1
redis==5.0.7
regex==2023.12.25
reportlab==3.6.8
requests==2.32.3
requests-file==1.5.1
rq==1.16.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
SecretStorage==3.3.1
service-identity==18.1.0
setproctitle==1.2.2
simplejson==3.17.6
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.5.0
ssh-import-id==5.11
systemd-python==234
tablib==3.5.0
tinycss2==1.1.1
tldextract==3.1.2
tomli==2.0.1
tornado==6.1
tqdm==4.66.4
translate-toolkit==3.6.0
Twisted==22.1.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
Unidecode==1.3.3
urllib3==2.2.2
validators==0.33.0
vobject==0.9.6.1
wadllib==1.3.6
webencodings==0.5.1
websockets==9.1
xdg==5
xkit==0.0.0
xlrd==1.2.0
yt-dlp==2022.4.8
zipp==1.0.0
zope.interface==5.4.0
